tests/wpt/web-platform-tests/app-history/navigate/reload-service-worker-fetch-event.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>

<!-- Keep an eye on https://github.com/whatwg/fetch/issues/1280 -->

<script>
const worker = "resources/fetch-event-test-worker.js";

promise_test(async t => {
  const scope = "resources/service-worker-page.html?reload-service-worker-fetch-event";

  const reg = await service_worker_unregister_and_register(t, worker, scope);

  await wait_for_state(t, reg.installing, "activated");
  const frame = await with_iframe(scope);
  assert_equals(frame.contentDocument.body.textContent, "method = GET, isReloadNavigation = false");

  frame.contentWindow.appHistory.reload();
  await new Promise(resolve => frame.addEventListener("load", resolve, { once: true }));
  assert_equals(frame.contentDocument.body.textContent, "method = GET, isReloadNavigation = true");

  frame.remove();
  await reg.unregister();
}, "reload() appears as a reload to service worker fetch event handlers");
</script>
